+++ tests/unit/base/test_utils.py
@@ -16,7 +16,6 @@ from rest_framework.test import APITestCase, APIClient
 from base.utils import RandomFileName, send_slack_notification, is_user_a_staff
 from challenges.models import Challenge, ChallengePhase
 from hosts.models import ChallengeHostTeam
-from hosts.utils import is_user_a_staff_or_host
 from jobs.models import Submission
 from participants.models import Participant, ParticipantTeam
 
@@ -158,62 +157,3 @@ class TestUserIsStaff(BaseAPITestClass):
         self.user.is_staff = False
         self.user.save()
         self.assertFalse(is_user_a_staff(self.user))
-
-
-class TestUserisStafforHost(BaseAPITestClass):
-    def setUp(self):
-        self.user = User.objects.create(
-            username="someuser1",
-            email="user@test.com",
-            password="secret_password",
-            is_staff=True,
-        )
-
-        self.user2 = User.objects.create(
-            username="someuser2",
-            email="user2@test.com",
-            password="secret_password2",
-            is_staff=False,
-        )
-
-        EmailAddress.objects.create(
-            user=self.user, email="user@test.com", primary=True, verified=True
-        )
-
-        EmailAddress.objects.create(
-            user=self.user2, email="user2@test.com", primary=True, verified=True
-        )
-
-        self.challenge_host_team = ChallengeHostTeam.objects.create(
-            team_name="Test Challenge Host Team", created_by=self.user
-        )
-
-        self.challenge = Challenge.objects.create(
-            title="Test Challenge",
-            description="Description for test challenge",
-            terms_and_conditions="Terms and conditions for test challenge",
-            submission_guidelines="Submission guidelines for test challenge",
-            creator=self.challenge_host_team,
-            start_date=timezone.now() - timedelta(days=2),
-            end_date=timezone.now() + timedelta(days=1),
-            published=False,
-            enable_forum=True,
-            anonymous_leaderboard=False,
-        )
-
-    def test_if_user_is_staff(self):
-        self.assertTrue(is_user_a_staff_or_host(self.user, self.challenge.pk))
-
-    def test_if_user_is_not_staff(self):
-        self.assertFalse(is_user_a_staff_or_host(self.user2, self.challenge.pk))
-
-    def test_if_user_is_host(self):
-        self.user.is_staff = False
-        self.user.save()
-        self.assertTrue(is_user_a_staff_or_host(self.user, self.challenge.pk))
-
-    def test_if_user_is_not_host(self):
-        self.assertFalse(is_user_a_staff_or_host(self.user2, self.challenge.pk))
-
-    def test_if_user_is_not_host_or_staff(self):
-        self.assertFalse(is_user_a_staff_or_host(self.user2, self.challenge.pk))
